Add public accessors for types... (XCode 6b4)
This is the first step in getting Nimble ready for XCode 6 beta 4.

The following things are included in this:
- public accessors added
- removed the bridging header (not supported in Frameworks)
- forward declared NMBExpectation (was getting an error related to it)

I'm still getting an error related to:

Undefined symbols for architecture x86_64:
  "_OBJC_CLASS_$_NMBExpectation", referenced from:
      objc-class-ref in DSL-B83371164C24FC7F.o
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)

But I'm unfamiliar with what the problem is here.