NimbleAssertionHandler is now public.
This allows custom replacements of Nimble's assertion infrastructure to
use an alternative test running system. The default uses
NimbleXCTestHandler, which passes failed assertions to XCTest.

Renamed CurrentAssertionHandler to NimbleAssertionHandler.

Nimble is prefixed to be more explicit and avoid possible confusion.
It's usage is intended to be infrequent if used at all.