Add `to` to expectation
It's currently a computed property that merely returns `self`. We may
want to eventually use some other kind of `_ExpectationType`
(`Anticipation`?) to better direct autocomplete.

Note that `_ExpectationType` is currently an internal implementation and
should not be conformed to (it depends on being force-cast to
`Expectation`). The protocol, however, is required since you can't
extend specialized classes, structures, or enumerations directly.

Signed-off-by: Stephen Celis <stephen@stephencelis.com>