Breaking: async expectations to be less flaky.
This commit utilizes dispatch sources to perform all its asynchronous
polling / timeout behavior instead of relying on the run loop. Sources
is a lower-level than using NSRunLoop which gives us more control over
when timers and polling intervals are executed.

Also, nesting any combination of waitUntil and
expect(...).toEventually(...) is not allowed because all asynchronous
checking code manages the run loop, which makes it inheritantly
unreliable to use at the same time. Nimble will now throw an exception
when it detects this scenario, although it is up to the test writer to
ensure he/she doesn't manually manipulate the run loop while
asynchronous expectations are on-going.

Breaking change: waitUntil's closure argument is executed synchronously
instead of dispatching to asynchronously to the main thread.