Swiftlint warnings breaking build
When building the project in Carthage with swiftlint installed, there are 11 warnings issued that trip the build.

This PR solves all but one by amending the code to correct the offending issues.

The last one is fixed by adding a specific disable. This was choosen due to lack of deep knowledge of the finner details of the codebase and the fact that the affected code is underway changes on future PR that are currently WIP.