Fix race condition in waitUntil causing RunLoop to freeze
This fixes a race condition that resulted in the main RunLoop locking up, halting all tests indefinitely.

This occured sometimes when `done()` was called asynchronously on a background thread.
We are checking if the result for the test is complete (ie. `done()` has been called) on line 258, then running the `RunLoop`. If the background thread marks the test complete between the check on line 258 and the running of the `RunLoop` on line 260, it stops the `RunLoop` **prior** to the `RunLoop` being started instead of after. This means the `RunLoop` is never stopped and the test hangs forever. This ensures that the completion of the test and the stopping of the `RunLoop` is called serially on the main thread, which prevents the race condition.

Unit test is included to verify the fix.