explicitly set supported platforms
this fixes building with Carthage on a DTK where the tvOS SDK is currently
missing. If the SDK is present, Xcode correctly infers the supported platforms
for tvOS, but if not it sets the supported platform to macOS. This makes
Carthage think that the tvOS target is a valid build target on the DTK even
when explicitly telling it to only build macOS targets.