Allow using async/await in expect
This changes Expectation into a protocol, and converts the old Expectation struct into SyncExpectation and AsyncExpectation.
Expression is still synchronous, and toEventually is dropped for AsyncExpectations.
Additionally, renamed the older Async methods to Polling to try to reduce confusion as to what toEventually actually does.

Updated documentation, both to explain what toEventually actually does, and to elaborate that Nimble supports async/await.