Introduce the require dsl. For when you need the assertion to pass before continuing (#1103)
* Introduce the require-dsl. expect, but it returns the result of the expression

As implied, this is basically a copy-paste of the functionality of expect. It's not complete, and it's not well-tested. But it gets the idea across.
Future work is cleaning this up, backfilling tests for features (such as the unwrap function), and adding support for polling requirements (i.e. toEventually).

Other than functionality, require also files the errorThrown issue type with XCTest, whereas expect files the assertionFailed issue type. This has minor differences that are mostly semantics.

Also, in addition to the require dsl, this also adds unwrap, which is a shorthand for `require(...).toNot(beNil())`.

* Backfill some tests for requirement

Support async, but not yet polling.

* CustomNSError is only defined in Foundation

* Support polling using require

Oh, gosh. This is terrible. I think a cleaner way to do this would be to change MatcherResult to also store the most recent value.
But that would be a breaking change, and might require updating existing matchers, so... no.

* require should allow you to discard the result

Rename requireAsync to requirea - matching the example set forth by expecta.
Adds requires, which is requirea, but for specifying SyncRequirement.

* Document require in the README

Update the table of contents with doctoc.

* Remove AssertionHandler.require., require matcher failures will be recorded as test assertion failures, which is what they are anyway.

* backfill tests for pollUnwrap

Also actually mark the non-polling require dsl as discardableResult

* Mark Expectation.onFailure as deprecated in favor of the require dsl

* Rename RequirementError to RequireError

Fixed test failure on linux

* Allow passing a custom error to be thrown by require

* Update readme TOC