Reimplement polling expectations to make them more reliable under heavy system loads (#1199)
* Reimplement polling expectations to make them more reliable under heavy system loads

This trades wall-clock time accuracy for reliability. It's nowhere near as important for a polling expectation to finish within the specified duration as it is to actually poll as expected.

Basically, under heavy system loads, the actual gap between polling attempts increases. Under sufficiently heavy loads, with a sufficiently low timeout, we might never have polled in the first place.

Instead, this calculate how many times we should poll by dividing the timeout by the interval. We then poll exactly that many times, sleeping for however long the polling interval is.

If polling interval is ever 0, then we set it to 1 nanosecond.

WaitUntil is also re-implemented in both async and sync, but that's because the older infrastructure for doing polling was replaced. Starting in Swift 6.3, we will emit a warning when WaitUntil is used in Swift Testing.

This change also removes the ability to detect a blocked runloop, so Nimble will no longer detect for that. In my opinion, this is a worthwhile tradeoff, as blocked runloop was a source of test flakiness.

* Bump xcode version, drop < swift 6.0

* Update workflows to specify xcode versions including swift 6

Fix build errors on older xcodes/linux

* Fix swiftlint errors

* Fix async waitUntil waiting until whenever the closure finishes before returning

* Remove unnecessary CoreFoundation import

* Update the test watch simulator

* Check for more CF imports, and guard it where necessary

* Mitigate AsyncAwaitTests.testWaitUntilDoesNotCompleteBeforeRunLoopIsWaiting flakes some more